Extract `RuleDefinition` as the base identity for rules
