+++ lib/rubydex.rb
@@ -20,6 +20,7 @@ require "rubydex/config"
 require "rubydex/location"
 require "rubydex/comment"
 require "rubydex/severity"
+require "rubydex/rule_definition"
 require "rubydex/related_information"
 require "rubydex/diagnostic"
 require "rubydex/keyword"
