+++ lib/rubydex/rule_definition.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+module Rubydex
+  # The identity of a rule. We have different mechanisms for collecting diagnostics, but they all share the same
+  # fundamental identity: a unique name and a default severity that may be overridden by the user's configuration.
+  #
+  # By keeping this identity concept unified, we are able to treat rules consistently regardless of whether they were
+  # collected by the graph during the analysis or by the linter afterwards.
+  #
+  # @abstract
+  class RuleDefinition
+    class << self
+      #: () -> String
+      def rule_name
+        name #: as !nil
+          .split("::").last #: as !nil
+      end
+
+      # @abstract
+      #: () -> singleton(Severity::Base)
+      def default_severity
+        raise NotImplementedError, "Subclasses must implement the default_severity method"
+      end
+    end
+  end
+end
